Update weekly report, hour reports

+++ entropy-compressor/src/main/java/algo/General.java
@@ -9,10 +9,10 @@ import datastructs.MyQueue;
 public final class General {
 
     public static final int CODEVALUEBITS = 16;
-    public static final int TOPVALUE = (1 << CODEVALUEBITS) - 1;
-    public static final int FIRSTQUARTER = TOPVALUE / 4 + 1;
-    public static final int HALF = FIRSTQUARTER * 2;
-    public static final int THIRDQUARTER = FIRSTQUARTER * 3;
+    public static final long TOPVALUE = (1L << CODEVALUEBITS) - 1;
+    public static final long FIRSTQUARTER = TOPVALUE / 4 + 1;
+    public static final long HALF = FIRSTQUARTER * 2;
+    public static final long THIRDQUARTER = FIRSTQUARTER * 3;
     /**
      * Global maximum amount of symbols supported. Most datastructs are inteded
      * to work with characters as ints, so this effectively translates to the
